feat(UI): move model context in separate page
