+++ ScriptBeeClient/src/app/pages/projects/project-details/model/context-model-page/context-model-page.component.html
@@ -1,0 +1,7 @@
+@if (instanceId()) {
+  <app-currently-loaded-models [project]="project()" [instanceId]="instanceId()!" />
+
+  <mat-divider />
+
+  <app-project-context [projectId]="project().id" [instanceId]="instanceId()!" />
+}
