+++ ScriptBeeClient/src/app/pages/projects/project-details/model/context-model-page/project-context/project-context.component.html
@@ -1,7 +1,7 @@
 <div class="project-model-section-title project-model-content-after-divider">Project Context</div>
 
 @if (projectContextResource.isLoading()) {
-  <app-centered-spinner text="Loading Context"></app-centered-spinner>
+  <app-loading-progress-bar title="Loading Context..."></app-loading-progress-bar>
 } @else if (projectContextResourceError()) {
   <app-error-state [error]="projectContextResourceError()!" (retry)="projectContextResource.reload()"></app-error-state>
 } @else {
@@ -10,15 +10,15 @@
   }
 
   @if (isClearContextLoading()) {
-    <app-centered-spinner text="Clearing Context"></app-centered-spinner>
+    <app-loading-progress-bar title="Clearing Context..."></app-loading-progress-bar>
   }
 
   @if (isReloadContextLoading()) {
-    <app-centered-spinner text="Reloading Context"></app-centered-spinner>
+    <app-loading-progress-bar title="Reloading Context..."></app-loading-progress-bar>
   }
 
   @if (isGenerateClassesLoading()) {
-    <app-centered-spinner text="Generating Classes"></app-centered-spinner>
+    <app-loading-progress-bar title="Generating Classes..."></app-loading-progress-bar>
   }
 
   <app-selectable-tree [data]="context()" [displayNameAccessor]="displayNameAccessor"></app-selectable-tree>
