+++ ScriptBeeClient/src/app/pages/projects/project-details/model/context-model-page/project-context/project-context.component.scss
