+++ ScriptBeeClient/src/app/pages/projects/project-details/model/project-model-page.component.scss
@@ -6,10 +6,6 @@
   place-content: center space-around;
   align-items: center;
   overflow: auto;
-}
-
-.project-details-card {
-  width: 75%;
   margin-top: 4%;
   margin-bottom: 4%;
 }
