+++ ScriptBeeClient/src/app/pages/projects/project-details/model/upload-model-page/upload-model-page.component.html
@@ -23,7 +23,7 @@
   <div class="warning keep-in-mind">Keep in mind: When uploading new models for a particular loader, the references to previous uploads will be deleted!</div>
 
   @if (isUploadLoading()) {
-    <app-centered-spinner text="Uploading Files"></app-centered-spinner>
+    <app-loading-progress-bar title="Uploading Files..."></app-loading-progress-bar>
   }
 
   <div class="upload-files-container">
